Switch all fatal/error print statements to the built in log mechanism
From now on all errors, warnings and fatal messages, should all go through
the standard log mechanism. This ensures that checking the log in automated
fashions in much simpler due to the use of standard "needles" that one can
search for, and also ensures that CBM error messages are visually distinct
from swupd, or any other tool, when there is combined log output.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>