bootman: Add RAII based CbmMappedFile
This simplifies the various paths used when dealing with an mmap()'d file,
including opening it, setting the pointer, and obtaining the size. It also
allows us to use the Nica autofree system by initialising a pointer with
CBM_MAPPED_FILE_INIT to a stack-local anonymous struct.

With this method, we can autofree() define a CbmMappedFile within the local
scope, and it will be automatically cleaned up for us as soon as it exits
that scope, simplifying any interaction with mmap'd files vastly.

This will be used to help replace the existing cbm_files_match function
and drop the dependency on OpenSSL.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>