tests: Add complete suite for cbm_files_match
This is just part of improving coverage within clr-boot-manager, and is
a critical requirement given that we have changed how we determine that
two files differ.

With that said, we still do have the test harness that makes heavy use of
cbm_files_match, and one can clearly see that by deliberately inverting
logic in paths within cbm_files_match, the harness tests then completely
fail.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>