syslinux: Fix some potential memory access issues
Here we simplify the use of default_text by only using it if it is actually
set, otherwise defaulting to "". Likewise we should only try to compare
the kernel path with the default kernel, if the provided kernel is not
actually NULL, leading to a null pointer dereference.

Whilst these issues haven't appeared thus far, it is better to be safe
rather than sorry.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>