bootman: Append kernel cmdline from globally defined kernel cmdline file
When updating the kernel, we can now optionally take additional content
from the /etc/kernel/cmdline file to append to the generated cmdline
options.

This allows users to add specific changes on their own hosts, restoring
flexibility and control. To be most useful this option would also be
improved by a "--force" option, to rebuild all the config files.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>