Always attempt installation of kernel, regardless of being installed
The prior method was very black or white, in which we would either install
a kernel, or continue. However, as we add more and more options into the
tool, it becomes clear that we also need to dynamically update the target
configuration files if one of the user defined configuration files changes.

To achieve this, we'll always attempt the install, which is currently
inefficient. However, future changes will focus on making this only modify
bits that actually need changing/propogating, which will also allow a more
granular approach to repairs.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>