bootloaders/systemd: Only write the text files when they're different
Given the previous 2 changes to ensure that we can dynamically update
the kernels and configs, we must ensure that we minimize all writes by
clr-boot-manager where possible.

With this change, we only write out the files when the proposed new
configuration differs in some way with the current configuration. In
all other instances, we just deem the configuration "written", and keep
the ESP happy.

Additionally, we replace the existing file_get_text method with something
more efficient and less virulent to valgrind by reusing our simplistic
mmap wrapper to just dup the whole data block in place.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>