bootman: Append kernel cmdline from globally defined kernel cmdline file
If the global cmdline exists, as defined by the cascade policy, we then
append this cmdline to the predefined kernel's upstream cmdline. This
ensures the complete cmdline is propogated to the bootloader mechanisms
in the least intrusive fashion.

We always pass the root= parameter first, followed by the kernel's own
parameters, and then /etc/kernel/cmdline, followed by the system glob
order of /etc/kernel/cmdline.d/*.conf.

Thus, it is possible to predict the complete cmdline order always, and
assertively validate it.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>