bootman: Only update the kernel blob when it's different
As clr-boot-manager may be called at any time, and we now do not bail
if the kernel is installed, we should only update the assets that changed.
In this instance, we'll write the kernel blob if the target doesn't exist
or if the source and target are now different.

This makes the update operation far cheaper, and respects the nature of the
ESP. This change is a critical part of issue #31.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>