bootloader/systemd: Only write the text files when they're different Given the previous changes to ensure that we can dynamically update the kernels and configs, we must ensure that we minimize all writes by clr-boot-manager where possible.
With this change, we only write out the files when the proposed new
configuration differs in some way with the current configuration. In
all other instances, we just deem the configuration "written", and keep
the ESP happy.

Additionally, the method for fallback timeout writing was altered to
reduce code duplication and comply with the policy of only writing the
files when they have actually changed.

This finalizes the systemd changes required for issue #31.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>