bootloaders/syslinux:
As the final part of issue #31, we now ensure that the syslinux bootloader
implementation also complies with the systemd class, by only writing the
configuration change when it is either different from the target syslinux
configuration, or the the target configuration doesn't yet exist.

We also ensure we minimise the cbm_sync() to cases when the write actually
took place for new configurations.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>