travis: Enable C11 + GCC5 by moving to trusty
Currently the only way to support the newer toolchain and have C11
support is by enabling Trusty. Admittedly this will make the builds
take longer, however our use of sudo already limited us to a VM on
the Google Compute Engine, so it's not too bad.

With this change we move to GCC 5 with it's complete C11 support, as
opposed to the previous GCC 4.8.0, which was even broken in various
places when it came to C99.

Having a closer matching toolchain version should ensure that any
compilation related issues are easier to debug, and ensure that the
CI environment is closer to that of the target deployments. This
should also resolve some of the very annoying ancient compiler warnings
that aren't valid with complete C99/C11 implementations.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>