bootloaders/syslinux: Add CbmWriter and initrd support
In keeping with changes to the systemd-class implementation, we now
drastically simplify much of the syslinux implementation, using cleaner
and efficient APIs for string management. This has a net improvement on
the coverage in the codebase too, removing many of the !malloc, fail
style routines.

Additionally, we can take the lead of the systemd-class implementation and
very easily plug in initrd support for the syslinux implementation.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>