lib: Add method to probe API for discovering LUKS parents
In order to support common deployment patterns, we also need to be able to
find out if a root system belongs to a LUKS disk. This change adds support
for detecting an LVM on LUKS partition, by checking that the root device is
actually a device mapper node, and has a depth of 2 slave systems.
Lastly, we then check that the TYPE value of the probe is crypto_LUKS, which
in turn is a constant string defined internally within util-linux, when
the LUKS container has been successfully detected via magic headers and
type detection.

This will be used in the next change to automatically insert the correct
kernel parameters into cmdline entries, enabling automatic configuration
of LUKS root UUID.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>