lib: Wrap the mount helper functions in the system vtable
Currently clr-boot-manager will (correctly) detect a mounted ESP on the
host within the test suite, causing failures as it tries to write outside
of the tree.

This change wraps the is_mounted, and get_mountpoint_for_device functions
in the system vfunc table, so that they can be forcibly overriden within
the test suite.

As the error paramater was never used in is_mounted, we drop it from the
codebase. Given that we may end up attempting to mount again, this will
then fail further along regardless if it was already mounted. We only
want to know that it definitely IS mounted.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>