Add support for automated building within docker environment.
In order to assist in building up test coverage, we need to do so in a way
that allows to test dangerous changes locally. Emulating an environment that
has no /sys or /dev is absurdly difficult.

Instead, we add a method to enable building (reproducibly) within a Docker
image based on Clear Linux. The build will take a full snapshot of the
current tree from a *read-only* bind mount into the target tree. At this
point it will attempt all parts of the normal Travis build steps, and will
ultimately emit the contents of the build log should failures be encountered.

Additionally we have the added benefit of working with a modern toolchain,
as opposed to the very old offerings on Travis.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>