lib: Add mock encapsulation for all /sys and /dev paths
In order to facilitate mocking, we must stop hardcoding /sys and /dev
paths. Unfortunately, these are the *right* paths to use in real life.
However, during testing, we most certainly do not want to use the host-side
virtual filesystems.

As such we change our assumptions and ensure /sys and /dev are baked in
constants in the test suite and internal to CBM, allowing us to now populate
fake vfs trees. Additionally, this will allow us to turn on and disable the
"UEFI support" in clr-boot-manager.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>