tests: Begin work on new skeletal UEFI test suite
This mimics the check-legacy test almost identically, but taking the
inverse road in certain areas to cater specifically for UEFI. Notably,
it also ensures that Legacy Boot is not detected.

This change will allow us to begin fleshing the main test suites out
in parallel to ensure they do not drift.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>