Replace all critical uses of asprintf with string_printf
This will help in recovering the coverage numbers that have been lost to
untestable malloc-failure codepaths, by continuing to do exactly the same
thing as before: if malloc-fail, abort().

This also makes the code far simpler and more pleasant to navigate and
removes a huge number of the abort calls from the codebase. Some still
exist but they are both minimal and obvious.

This small, seemingly trivial change, restores almost 2% coverage in the
codebase testing.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>