bootloaders: Introduce a very trivial capabilities system
This is used to describe in an obvious sense the purpose and application
for a given bootloader implementation. For now we'll use this to determine
whether a selected bootloader is being used in UEFI mode or not, and in the
future we can expand on this to dynamically select the correct bootloader
based on the system topology.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>