bootloader/systemd-class: Add support for secure boot
This change reorganises how we achieve the pairing of files, by literally
implementing a source->target map of files that will need copying by the
systemd class implementation. As such it makes it extensible.

The main gain for this change, is that when we detect the shim stage1
secure boot loader available, we'll automatically make use of it and
install it as the target BOOTX64.EFI, with a fallback loader$SUFFIX.efi.

The name can be changed (i.e. to "grub") for older tests, however this
behaviour is currently dependent on patches to both systemd-boot and
shim, which will be upstreamed in due course.

Lastly, we will be able to easily extend this functionality in future to
support automatic installation of fwupd to the ESP for firmware updates,
and ensure that fwupd itself is kept up to date.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>