bootman: Automatically purge associated header tree
Many distributions permit building out of tree modules against the headers
for each kernel. To allow this to happen in a safely managed fashion, CBM
must be able to remove these paths from disk and allow the software
deployment mechanism to mark these paths as resident.

The net effect is that one can roll back to an older kernel, and still have
the correct headers available for kernel module compilation.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>