bootman: Collapse double slashed boot directory where possible
In the event that the boot dir exists, we can realpath it to collapse our
returned path to remove any double slashes which in turn would've stopped
the lookup function working for cbm_is_mounted, when determining if the ESP
is already mounted or not.

This helps, in part, issue #54.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>