bootman: Never match last_booted to running kernel
While an update currently performs just fine with clr-boot-manager, a
subsequent call to clr-boot-manager update, from a rebooted, updated system,
using tip & current, will result in the last kernel being garbage collected.

This is because last_good == running == tip, thus anything else is now
considered for garbage collection. With this simple change, we ensure that
last_running never matches our currently running kernel, as it is OK for
the last running kernel to be unknown. If it is known, then it truly is the
"newest" last kernel that isn't the one we're on now.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>