bootloaders/syslinux: Avoid duplicated entries in output
Due to our repair vs 1:1 "is installed" method changes, it is possible that
a kernel may be asked to be installed more than one time. As such we modify
the syslinux implementation to match that of GRUB2, and ensure that the
kernels being added to the list are all unique.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>