bootman: Remove legacy kernel blobs from bootdir
Since we switched to legacy vs UEFI namespacing, we only removed the target
path for the internal kernel removal code. This means that non UEFI systems
are being left with old blobs, unmanaged, on the boot partition.

This change ensures we always remove excess blobs for legacy booting systems
and not wasting space in that boot partition/dir.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>