kernel: Allow lack of modules to be non-fatal
The lack of kernel modules during install isn't necessarily fatal, and
some kernel configurations might be without modules entirely. Notably,
given that kernel modules are to be marked as resident on disk, even if
we have multiple packages compromising a kernel and separate modules, those
old modules will be removed at the time of the kernel change, and at no
other time, ensuring an atomic update.

This resolves #67.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>