Enable listing of available kernels and setting a default
Add the cli command list-kernels which will display the currently
available kernels that can be booted on the system and highlight the
kernel which will be used for the next boot.

Also add the cli command set-kernel which will set the given kernel as
the default to be used for booting (until the next update or
set-kernel is run).